moving artifactory
